+++ src/main/java/net/sf/jabref/gui/JabRefFrame.java
@@ -105,6 +105,7 @@ import net.sf.jabref.gui.protectedterms.ProtectedTermsDialog;
 import net.sf.jabref.gui.push.PushToApplicationButton;
 import net.sf.jabref.gui.push.PushToApplications;
 import net.sf.jabref.gui.search.GlobalSearchBar;
+import net.sf.jabref.gui.specialfields.SpecialFieldDropDown;
 import net.sf.jabref.gui.util.FocusRequester;
 import net.sf.jabref.gui.util.WindowLocation;
 import net.sf.jabref.gui.worker.MarkEntriesAction;
@@ -650,7 +651,7 @@ public class JabRefFrame extends JFrame implements OutputPrinter {
             } else {
                 String content = "";
                 SearchQuery currentSearchQuery = currentBasePanel.getCurrentSearchQuery();
-                if (currentSearchQuery != null && !currentSearchQuery.getQuery().trim().isEmpty()) {
+                if ((currentSearchQuery != null) && !currentSearchQuery.getQuery().trim().isEmpty()) {
                     content = currentSearchQuery.getQuery();
                 }
                 globalSearchBar.setSearchTerm(content, true);
@@ -1431,8 +1432,7 @@ public class JabRefFrame extends JFrame implements OutputPrinter {
         tlb.addAction(unmark);
         if (Globals.prefs.getBoolean(JabRefPreferences.SPECIALFIELDSENABLED)) {
             if (Globals.prefs.getBoolean(JabRefPreferences.SHOWCOLUMN_RANKING)) {
-                JButton button = net.sf.jabref.specialfields.SpecialFieldDropDown
-                        .generateSpecialFieldButtonWithDropDown(Rank.getInstance(), this);
+                JButton button = SpecialFieldDropDown.generateSpecialFieldButtonWithDropDown(Rank.getInstance(), this);
                 tlb.add(button);
                 specialFieldButtons.add(button);
             }
@@ -1443,8 +1443,8 @@ public class JabRefFrame extends JFrame implements OutputPrinter {
                 tlb.addAction(toggleQualityAssured);
             }
             if (Globals.prefs.getBoolean(JabRefPreferences.SHOWCOLUMN_PRIORITY)) {
-                JButton button = net.sf.jabref.specialfields.SpecialFieldDropDown
-                        .generateSpecialFieldButtonWithDropDown(Priority.getInstance(), this);
+                JButton button = SpecialFieldDropDown.generateSpecialFieldButtonWithDropDown(Priority.getInstance(),
+                        this);
                 tlb.add(button);
                 specialFieldButtons.add(button);
             }
@@ -1452,8 +1452,8 @@ public class JabRefFrame extends JFrame implements OutputPrinter {
                 tlb.addAction(togglePrinted);
             }
             if (Globals.prefs.getBoolean(JabRefPreferences.SHOWCOLUMN_READ)) {
-                JButton button = net.sf.jabref.specialfields.SpecialFieldDropDown
-                        .generateSpecialFieldButtonWithDropDown(ReadStatus.getInstance(), this);
+                JButton button = SpecialFieldDropDown.generateSpecialFieldButtonWithDropDown(ReadStatus.getInstance(),
+                        this);
                 tlb.add(button);
                 specialFieldButtons.add(button);
             }
