+++ src/main/java/net/sf/jabref/gui/specialfields/SpecialFieldDropDown.java
@@ -1,4 +1,4 @@
-package net.sf.jabref.specialfields;
+package net.sf.jabref.gui.specialfields;
 
 import java.awt.Dimension;
 import java.awt.Insets;
@@ -11,7 +11,9 @@ import javax.swing.JPopupMenu;
 
 import net.sf.jabref.gui.BasePanel;
 import net.sf.jabref.gui.JabRefFrame;
+import net.sf.jabref.logic.specialfields.SpecialField;
 import net.sf.jabref.logic.util.OS;
+import net.sf.jabref.specialfields.SpecialFieldValue;
 
 import com.jgoodies.looks.HeaderStyle;
 import com.jgoodies.looks.Options;
