+++ src/main/java/net/sf/jabref/specialfields/Rank.java
@@ -5,6 +5,7 @@ import java.util.List;
 
 import net.sf.jabref.gui.IconTheme;
 import net.sf.jabref.logic.l10n.Localization;
+import net.sf.jabref.logic.specialfields.SpecialField;
 import net.sf.jabref.model.entry.SpecialFields;
 
 public class Rank extends SpecialField {
