+++ src/main/java/net/sf/jabref/specialfields/SpecialFieldValue.java
@@ -6,6 +6,9 @@ import javax.swing.Icon;
 import javax.swing.JLabel;
 
 import net.sf.jabref.gui.JabRefFrame;
+import net.sf.jabref.gui.specialfields.SpecialFieldAction;
+import net.sf.jabref.gui.specialfields.SpecialFieldMenuAction;
+import net.sf.jabref.logic.specialfields.SpecialField;
 
 public class SpecialFieldValue {
